This is a To Do App
This code defines a ToDoItem class with properties for the title, due date, and completion status. It also defines a ToDoList class with methods for adding items, marking items as completed, and getting items that are due within a certain number of days.

The AppDelegate class contains a ToDoList instance and methods for adding items, marking items as completed, and getting items that are due within a certain number of days. It also includes code for scheduling reminders for items using the UserNotifications framework.

You can then use the addItem, markItemCompleted, and getItemsWithinDays methods in your app to manage the to-do list and display items to the user. For example, you could create a table view controller that displays a list of items due within the next 7 days, and allow the user to mark items as completed by swiping left on the table view cell. You could also schedule a local notification for each item using the scheduleReminder method.